+++ src/zimg/unresize/bilinear.cpp
@@ -61,19 +61,19 @@ TridiagonalLU<T> tridiagonal_decompose(const RowMatrix<T> &m)
  */
 RowMatrix<double> bilinear_weights(unsigned in, unsigned out, double shift)
 {
-	RowMatrix<double> m{ (size_t)out, (size_t)in };
+	RowMatrix<double> m{ out, in };
 
 	// Position of outermost samples on input grid.
 	double leftmost = 0.5 + shift;
 	double rightmost = in - 0.5 + shift;
 
 	// Indices corresponding to the samples stored at leftmost and rightmost.
-	unsigned leftmost_idx = (unsigned)std::min(std::max(std::floor(leftmost), 0.0), (double)in - 1.0);
-	unsigned rightmost_idx = (unsigned)std::min(std::max(std::floor(rightmost), 0.0), (double)in - 1.0);
+	unsigned leftmost_idx = static_cast<unsigned>(std::min(std::max(std::floor(leftmost), 0.0), in - 1.0));
+	unsigned rightmost_idx = static_cast<unsigned>(std::min(std::max(std::floor(rightmost), 0.0), in - 1.0));
 
 	for (unsigned i = 0; i < out; ++i) {
 		// Position of output sample on input grid.
-		double position = (i + 0.5) * (double)in / (double)out;
+		double position = (i + 0.5) * in / out;
 
 		// For samples outside the input range, mirror the nearest input pixel.
 		if (position <= leftmost) {
@@ -82,7 +82,7 @@ RowMatrix<double> bilinear_weights(unsigned in, unsigned out, double shift)
 			m[i][rightmost_idx] = 1.0;
 		} else {
 			// Index of nearest input pixels to output position.
-			unsigned left_idx = (unsigned)std::floor(position - leftmost);
+			unsigned left_idx = static_cast<unsigned>(std::floor(position - leftmost));
 			unsigned right_idx = left_idx + 1;
 
 			// Distance between output position and left input.
@@ -107,7 +107,7 @@ BilinearContext create_bilinear_context(unsigned in, unsigned out, double shift)
 	BilinearContext ctx;
 
 	// Map output shift to input shift.
-	RowMatrix<double> m = bilinear_weights(in, out, -shift * (double)in / (double)out);
+	RowMatrix<double> m = bilinear_weights(in, out, -shift * in / out);
 	RowMatrix<double> transpose_m = ~m;
 	RowMatrix<double> pinv_m = transpose_m * m;
 	TridiagonalLU<double> lu = tridiagonal_decompose(pinv_m);
@@ -122,28 +122,28 @@ BilinearContext create_bilinear_context(unsigned in, unsigned out, double shift)
 	for (size_t i = 0; i < rows; ++i) {
 		rowsize = std::max(transpose_m.row_right(i) - transpose_m.row_left(i), rowsize);
 	}
-	size_t rowstride = (unsigned)ceil_n(rowsize, AlignmentOf<float>::value);
+	size_t rowstride = ceil_n(rowsize, AlignmentOf<float>::value);
 
 	ctx.matrix_coefficients.resize(rowstride * rows);
 	ctx.matrix_row_offsets.resize(rows);
-	ctx.matrix_row_size = (unsigned)rowsize;
-	ctx.matrix_row_stride = (unsigned)rowstride;
+	ctx.matrix_row_size = static_cast<unsigned>(rowsize);
+	ctx.matrix_row_stride = static_cast<unsigned>(rowstride);
 	for (size_t i = 0; i < rows; ++i) {
 		size_t left = std::min(transpose_m.row_left(i), cols - rowsize);
 
 		for (size_t j = 0; j < transpose_m.row_right(i) - left; ++j) {
-			ctx.matrix_coefficients[i * rowstride + j] = (float)transpose_m[i][left + j];
+			ctx.matrix_coefficients[i * rowstride + j] = static_cast<float>(transpose_m[i][left + j]);
 		}
-		ctx.matrix_row_offsets[i] = (unsigned)left;
+		ctx.matrix_row_offsets[i] = static_cast<unsigned>(left);
 	}
 
 	ctx.lu_c.resize(rows);
 	ctx.lu_l.resize(rows);
 	ctx.lu_u.resize(rows);
 	for (size_t i = 0; i < rows; ++i) {
-		ctx.lu_c[i] = (float)lu.c[i];
-		ctx.lu_l[i] = (float)(1.0 / (lu.l[i] + epsilon<float>())); // Pre-invert this value, as it is used in division.
-		ctx.lu_u[i] = (float)lu.u[i];
+		ctx.lu_c[i] = static_cast<float>(lu.c[i]);
+		ctx.lu_l[i] = static_cast<float>((1.0 / (lu.l[i] + epsilon<float>()))); // Pre-invert this value, as it is used in division.
+		ctx.lu_u[i] = static_cast<float>(lu.u[i]);
 	}
 
 	return ctx;
